Move legacy console code to IAccessible and only enable UIA when available.
