+++ source/NVDAObjects/IAccessible/winConsoleLegacy.py
@@ -1,23 +1,23 @@
-#NVDAObjects/WinConsole.py
+#NVDAObjects/IAccessible/winConsoleLegacy.py
 #A part of NonVisual Desktop Access (NVDA)
 #This file is covered by the GNU General Public License.
 #See the file COPYING for more details.
-#Copyright (C) 2007-2012 NV Access Limited
+#Copyright (C) 2007-2019 NV Access Limited, Bill Dengler
 
-import winConsoleHandler
-from . import Window
+import winConsoleHandlerLegacy as winConsoleHandler
+from . import IAccessible
 from ..behaviors import Terminal, EditableTextWithoutAutoSelectDetection
 import api
 import core
 
-class WinConsole(Terminal, EditableTextWithoutAutoSelectDetection, Window):
+class winConsoleLegacy(Terminal, EditableTextWithoutAutoSelectDetection, IAccessible):
 	STABILIZE_DELAY = 0.03
 
 	def _get_TextInfo(self):
 		consoleObject=winConsoleHandler.consoleObject
 		if consoleObject and self.windowHandle == consoleObject.windowHandle:
-			return winConsoleHandler.WinConsoleTextInfo
-		return super(WinConsole,self).TextInfo
+			return winConsoleHandler.legacyConsoleTextInfo
+		return super(winConsoleLegacy,self).TextInfo
 
 	def event_becomeNavigatorObject(self, isFocus=False):
 		if winConsoleHandler.consoleObject is not self:
@@ -28,17 +28,17 @@ class WinConsole(Terminal, EditableTextWithoutAutoSelectDetection, Window):
 				# The user is returning to the focus object with object navigation.
 				# The focused console should always be monitored if possible.
 				self.startMonitoring()
-		super(WinConsole,self).event_becomeNavigatorObject(isFocus=isFocus)
+		super(winConsoleLegacy,self).event_becomeNavigatorObject(isFocus=isFocus)
 
 	def event_gainFocus(self):
 		if winConsoleHandler.consoleObject is not self:
 			if winConsoleHandler.consoleObject:
 				winConsoleHandler.disconnectConsole()
 			winConsoleHandler.connectConsole(self)
-		super(WinConsole, self).event_gainFocus()
+		super(winConsoleLegacy, self).event_gainFocus()
 
 	def event_loseFocus(self):
-		super(WinConsole, self).event_loseFocus()
+		super(winConsoleLegacy, self).event_loseFocus()
 		if winConsoleHandler.consoleObject is self:
 			winConsoleHandler.disconnectConsole()
 
@@ -49,7 +49,7 @@ class WinConsole(Terminal, EditableTextWithoutAutoSelectDetection, Window):
 		return winConsoleHandler.getConsoleVisibleLines()
 
 	def script_caret_backspaceCharacter(self, gesture):
-		super(WinConsole, self).script_caret_backspaceCharacter(gesture)
+		super(winConsoleLegacy, self).script_caret_backspaceCharacter(gesture)
 		# #2586: We use console update events for typed characters,
 		# so the typedCharacter event is never fired for the backspace key.
 		# Call it here so that speak typed words works as expected.
